requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
cryptography==41.0.7
redis==5.0.1
pyyaml==6.0.1
structlog==24.1.0
prometheus-client==0.19.0
python-multipart==0.0.6
httpx==0.25.2
tenacity==8.2.3
python-json-logger==2.0.7
bcrypt==4.1.1
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.2
aioredis==2.0.1
